Refactor float package as a standard go package practice
