+++ utils/float/float.go
@@ -1,4 +1,4 @@
-package utils
+package float
 
 import "math"
